fix compilation errors

+++ client/src-tauri/src/capture/gpu_frame.rs
@@ -47,7 +47,7 @@ impl GpuTextureHandle {
                 },
                 Usage: D3D11_USAGE_STAGING,
                 BindFlags: Default::default(),
-                CPUAccessFlags: D3D11_CPU_ACCESS_READ.0,
+                CPUAccessFlags: D3D11_CPU_ACCESS_READ.0 as u32,
                 MiscFlags: Default::default(),
             };
